show dacs and links in dataset table
